+++ pkg/local_object_storage/metabase/metadata.go
@@ -13,7 +13,6 @@ import (
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
-	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"github.com/nspcc-dev/neofs-sdk-go/version"
 	"go.etcd.io/bbolt"
 	"go.uber.org/zap"
@@ -46,7 +45,7 @@ func invalidMetaBucketKeyErr(key []byte, cause error) error {
 // PutMetadataForObject fills object meta-data indexes using bbolt transaction.
 // Transaction must be writable. Additional bucket for container's meta-data
 // may be created using {255, CID...} form as a key.
-func PutMetadataForObject(tx *bbolt.Tx, hdr object.Object, hasParent, phy bool) error {
+func PutMetadataForObject(tx *bbolt.Tx, hdr object.Object, phy bool) error {
 	metaBkt, err := tx.CreateBucketIfNotExists(metaBucketKey(hdr.GetContainerID()))
 	if err != nil {
 		return fmt.Errorf("create meta bucket for container: %w", err)
@@ -107,7 +106,7 @@ func PutMetadataForObject(tx *bbolt.Tx, hdr object.Object, hasParent, phy bool) 
 			return err
 		}
 	}
-	if !hasParent && hdr.Type() == object.TypeRegular {
+	if !hdr.HasParent() && hdr.Type() == object.TypeRegular {
 		if err = putPlainAttribute(metaBkt, &keyBuf, id, object.FilterRoot, binPropMarker); err != nil {
 			return err
 		}
@@ -170,26 +169,6 @@ func deleteMetadata(tx *bbolt.Tx, l *zap.Logger, cnr cid.ID, id oid.ID, isParent
 					return 0, err
 				}
 			}
-		case object.FilterOwnerID:
-			if len(attrV) == user.IDSize {
-				owner := user.ID(attrV)
-				delFKBTIndexItem(tx, namedBucketItem{
-					name: ownerBucketName(cnr, bktKeyBuff),
-					key:  []byte(owner.EncodeToString()),
-					val:  id[:],
-				})
-			} else {
-				l.Warn("unexpected object's owner index",
-					zap.Stringer("addr", oid.NewAddress(cnr, id)),
-					zap.Int("len", len(attrV)),
-				)
-			}
-		case object.FilterPayloadChecksum:
-			delListIndexItem(tx, namedBucketItem{
-				name: payloadHashBucketName(cnr, bktKeyBuff),
-				key:  attrV,
-				val:  id[:],
-			})
 		case object.FilterParentID:
 			if len(attrV) == oid.Size {
 				parent = oid.ID(attrV)
@@ -203,29 +182,9 @@ func deleteMetadata(tx *bbolt.Tx, l *zap.Logger, cnr cid.ID, id oid.ID, isParent
 				key:  attrV,
 				val:  id[:],
 			})
-		case object.FilterSplitID:
-			delListIndexItem(tx, namedBucketItem{
-				name: splitBucketName(cnr, bktKeyBuff),
-				key:  attrV,
-				val:  id[:],
-			})
-		case object.FilterFirstSplitObject:
-			delListIndexItem(tx, namedBucketItem{
-				name: firstObjectIDBucketName(cnr, bktKeyBuff),
-				key:  attrV,
-				val:  id[:],
-			})
 		case object.FilterPayloadSize:
 			size, _ = strconv.ParseUint(string(attrV), 10, 64)
 		default:
-			const systemFieldPrefix = "$Object:"
-			if !bytes.HasPrefix(attrK, []byte(systemFieldPrefix)) {
-				delFKBTIndexItem(tx, namedBucketItem{
-					name: attributeBucketName(cnr, kStr, bktKeyBuff),
-					key:  attrV,
-					val:  id[:],
-				})
-			}
 		}
 		kAttrID := make([]byte, len(kIDAttr)+attributeDelimiterLen)
 		kAttrID[0] = metaPrefixAttrIDPlain
@@ -264,14 +223,14 @@ func deleteMetadata(tx *bbolt.Tx, l *zap.Logger, cnr cid.ID, id oid.ID, isParent
 
 // Search selects up to count container's objects from the given container
 // matching the specified filters.
-func (db *DB) Search(cnr cid.ID, fs object.SearchFilters, fInt map[int]objectcore.ParsedIntFilter, attrs []string, cursor *objectcore.SearchCursor, count uint16) ([]client.SearchResultItem, []byte, error) {
+func (db *DB) Search(cnr cid.ID, fs []objectcore.SearchFilter, attrs []string, cursor *objectcore.SearchCursor, count uint16) ([]client.SearchResultItem, []byte, error) {
 	var res []client.SearchResultItem
 	var newCursor []byte
 	var err error
 	if len(fs) == 0 {
 		res, newCursor, err = db.searchUnfiltered(cnr, cursor, count)
 	} else {
-		res, newCursor, err = db.search(cnr, fs, fInt, attrs, cursor, count)
+		res, newCursor, err = db.search(cnr, fs, attrs, cursor, count)
 	}
 	if err != nil {
 		return nil, nil, err
@@ -279,12 +238,12 @@ func (db *DB) Search(cnr cid.ID, fs object.SearchFilters, fInt map[int]objectcor
 	return res, newCursor, nil
 }
 
-func (db *DB) search(cnr cid.ID, fs object.SearchFilters, fInt map[int]objectcore.ParsedIntFilter, attrs []string, cursor *objectcore.SearchCursor, count uint16) ([]client.SearchResultItem, []byte, error) {
+func (db *DB) search(cnr cid.ID, fs []objectcore.SearchFilter, attrs []string, cursor *objectcore.SearchCursor, count uint16) ([]client.SearchResultItem, []byte, error) {
 	var res []client.SearchResultItem
 	var newCursor []byte
 	err := db.boltDB.View(func(tx *bbolt.Tx) error {
 		var err error
-		res, newCursor, err = db.searchTx(tx, cnr, fs, fInt, attrs, cursor, count)
+		res, newCursor, err = db.searchTx(tx, cnr, fs, attrs, cursor, count)
 		return err
 	})
 	if err != nil {
@@ -293,7 +252,7 @@ func (db *DB) search(cnr cid.ID, fs object.SearchFilters, fInt map[int]objectcor
 	return res, newCursor, nil
 }
 
-func (db *DB) searchTx(tx *bbolt.Tx, cnr cid.ID, fs object.SearchFilters, fInt map[int]objectcore.ParsedIntFilter, attrs []string, cursor *objectcore.SearchCursor, count uint16) ([]client.SearchResultItem, []byte, error) {
+func (db *DB) searchTx(tx *bbolt.Tx, cnr cid.ID, fs []objectcore.SearchFilter, attrs []string, cursor *objectcore.SearchCursor, count uint16) ([]client.SearchResultItem, []byte, error) {
 	metaBkt := tx.Bucket(metaBucketKey(cnr))
 	if metaBkt == nil {
 		return nil, nil, nil
@@ -315,7 +274,7 @@ func (db *DB) searchTx(tx *bbolt.Tx, cnr cid.ID, fs object.SearchFilters, fInt m
 		return objectStatus(tx, oid.NewAddress(cnr, id), curEpoch) == 0
 	}
 	resHolder := objectcore.SearchResult{Objects: make([]client.SearchResultItem, 0, count)}
-	handleKV := objectcore.MetaDataKVHandler(&resHolder, attrSkr, gcCheck, fs, fInt, attrs, cursor, count)
+	handleKV := objectcore.MetaDataKVHandler(&resHolder, attrSkr, gcCheck, fs, attrs, cursor, count)
 
 	for ; bytes.HasPrefix(primKey, cursor.PrimaryKeysPrefix); primKey, _ = primCursor.Next() {
 		if !handleKV(primKey, nil) {